% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seer.R
\name{fit_models}
\alias{fit_models}
\title{Fit Seer Models}
\usage{
fit_models(obj)
}
\arguments{
\item{obj}{seer object}
}
\value{
list with model algo, index, and fit in each element
}
\description{
Fits all models on all training samples in seer object
}
