% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{get_final_model}
\alias{get_final_model}
\title{Get Final Model}
\usage{
get_final_model(obj)
}
\arguments{
\item{obj}{seer object}
}
\value{
nested data.frame with new model fit, forecasts, and forecast weight
}
\description{
Fits final models on full training sample based on selection criterion.
Forecasts on refit models
}
