% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecasts.R
\name{get_forecasts}
\alias{get_forecasts}
\title{Get Forecasts}
\usage{
get_forecasts(...)
}
\arguments{
\item{...}{list of named arguments}
}
\value{
tibble with forecast predictions
}
\description{
Wrapper function for making forecasts using forecast function
}
