% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_performance.R
\name{get_model_performance}
\alias{get_model_performance}
\title{Get Model Performance}
\usage{
get_model_performance(obj)
}
\arguments{
\item{predictions}{nested data.frame with predictions. output of
`get_model_predictions` function}

\item{actuals}{data.frame with only one column that is the actual y_var}
}
\value{
data.frame with ME, RMSE, MAE, MPE, and MAPE calculated for all model
  uid and sample combinations
}
\description{
Function calculates model performance on all uid, sample, and index
predictions then averages by uid and sample. Function uses forecast
`accuracy` function to calculate metrics. Used internally by `seer` function
}
