% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_predictions.R
\name{get_model_predictions}
\alias{get_model_predictions}
\title{Get Model Predictions}
\usage{
get_model_predictions(obj)
}
\arguments{
\item{obj}{seer object}

\item{fits}{tibble with uid, index, and fit. Result of fit_models function}
}
\value{
nested tibble with model predictions
}
\description{
Function to get all seer model predictions for train and validation samples
}
\details{
Returned nested tibble has column for model uid, sample, index, and nested
data column with rn (row number), predicted and any confidence levels columns
}
