% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_final_forecasts.R
\name{make_final_forecasts}
\alias{make_final_forecasts}
\title{Make Final Forecasts}
\usage{
make_final_forecasts(obj)
}
\arguments{
\item{obj}{seer object}
}
\value{
tibble with final forecasts. Adds index column to forecast ouput
  which starts at 1 plus the number of rows in the input dataset
}
\description{
Makes final forecasts based on models and forecasting weight as a result of
model selection criteria
}
\details{
Confidence intervals are calculated by taking widest interval
}
