% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model}
\alias{model}
\title{Create Forecast Model}
\usage{
model(algo, args = list(), uid = madutils::random_string("model"),
  desc = "")
}
\arguments{
\item{algo}{model algorithm string. ex - `auto.arima`}

\item{args}{list of model arguments. default is empty list}

\item{uid}{optional sting uid argument. default is random_string}

\item{desc}{optional description}
}
\value{
model class object
}
\description{
Constructor function to create a model class object for use with seer
function. Creates an model object after assertions on the algorithm input.
Used to standarize the model inputs
}
\examples{
model(algo = "auto.arima", args = list(), uid = "my-model", desc = "example description")
}
