% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{model_selection}
\alias{model_selection}
\title{Model Selection Constructor}
\usage{
model_selection(measure = "RMSE", n = 1, weights = "equal")
}
\arguments{
\item{measure}{accuracy metric to arrange and weight by. Choices are RMSE and
MAE. Default is RMSE}

\item{n}{top_n number of models to select. Default is 1}

\item{weights}{option to weight model forcasts by inverse of measure or
equally. Options are equal for equal weighting or weighted for inverse
weighting. Default is equal}
}
\value{
model selection object
}
\description{
Helper function to create final model selection criteria for forecasting
}
\examples{
# default
model_selection()

# weighted selection
model_selection(measure = "RMSE", n = 2, weights = "weighted")
}
