% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{models}
\alias{models}
\title{Forecasting Model Algorithms for mad4sight}
\format{A data frame with 24 rows and 5 variables:
 \describe{
  \item{algorithm}{model algorithm function name}
  \item{name}{algorithm name}
  \item{package}{R package}
  \item{class}{method class membership. can be more than 1} ...
  }}
\usage{
models
}
\description{
A dataset containing the avaialble model algorithms supported in mad4sight
 from the forecast package
}
\keyword{datasets}
