% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seer.R
\name{neophyte}
\alias{neophyte}
\title{Seer Constructor Function}
\usage{
neophyte(df, y_var, sampling, models, measure, confidence_levels, horizon,
  forecast_xreg, backend)
}
\arguments{
\item{df}{data.frame input}

\item{y_var}{column name of target variable}

\item{sampling}{sample constructor function. see `samples` function for
details}

\item{models}{list of model constructor functions. see `model` function for
details}

\item{measure}{string input of error metric to use for model performance
evaluation}

\item{confidence_levels}{numeric vector with one or two confidence levels
used in the forecast predictions}

\item{horizon}{forecast horizon}

\item{forecast_xreg}{optional foreacast covariate data}

\item{backend}{future backend mode. `sequential` creates single threaded
execution. `multisession` creates parrallel backend}

\item{x_vars}{optional column names of covariates}
}
\value{
seer class object
}
\description{
Creates seer class object
}
