% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{samples}
\alias{samples}
\title{Samples Constructor}
\usage{
samples(method, args = list())
}
\arguments{
\item{method}{sampling method}

\item{args}{list of sampling method arguments}
}
\value{
samples class object
}
\description{
Function to create samples class object. Validates inputs with supported
sampling functions in mad4sight. Used internally within `seer` function
}
\examples{
samples(method = "split", args = list(ratio = .5))
}
