% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seer.R
\name{seer}
\alias{seer}
\title{Seer function to create forecasts}
\usage{
seer(df, y_var, sampling = samples(method = "split", args = list(ratio =
  0.9)), models = list(model(algo = "auto.arima")),
  selection = model_selection(), confidence_levels = c(80, 95),
  horizon = 1, forecast_xreg = NULL, backend = "sequential",
  user = NULL, uid = madutils::random_string("seer"), desc = "")
}
\arguments{
\item{df}{data.frame input}

\item{y_var}{column name of target variable}

\item{sampling}{sample constructor function. see `samples` function for
details}

\item{models}{list of model constructor functions. see `model` function for
details}

\item{selection}{model_selection constructor function result. see
`model_selection` for details}

\item{confidence_levels}{numeric vector with one or two confidence levels
used in the forecast predictions}

\item{horizon}{forecast horizon}

\item{forecast_xreg}{optional foreacast covariate data}

\item{backend}{future backend mode. `sequential` creates single threaded
execution. `multisession` creates parrallel backend}

\item{user}{optional string input for user. defaults to sys.user}

\item{uid}{optional string for uid. defaults to random string}

\item{desc}{optional string input for seer description. defaults to empty
string}
}
\description{
Seer function to create forecasts
}
