% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{single}
\alias{single}
\title{Create Single Sample}
\usage{
single(df)
}
\arguments{
\item{df}{data.frame to sample}
}
\value{
list with train and validation named indices
}
\description{
Function to create a single training sample formatted for mad4sight use
}
\examples{
library(tibble)
df <- tibble(i = 1:20)
single(df)
}
