% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{slices}
\alias{slices}
\title{Create Time Slice Samples}
\usage{
slices(df, width, horizon, skip = 0, fixed_width = TRUE,
  label = "slice")
}
\arguments{
\item{df}{data.frame to sample from}

\item{width}{number of records per training sample}

\item{horizon}{number of recoreds per validation sample}

\item{skip}{number of samples skipped. default is 0}

\item{fixed_width}{logical option to keep all training samples the same
width. default is TRUE}

\item{label}{string label for samples. default is `slice`}
}
\value{
list with train and validation samples
}
\description{
Function to create time slice indices. Based on
\link[Caret]{createTimeSlices} from Caret package
}
\examples{
library(tibble)
df <- tibble(i = 1:20)
slices(df, width = 15, horizon = 1, skip = 0)
}
