% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{split}
\alias{split}
\title{Create Split Samples}
\usage{
split(df, ratio)
}
\arguments{
\item{df}{data.frame to sample}

\item{ratio}{ratio of data to use for train sample. 1-ratio used for validation}
}
\value{
list with train and validation samples
}
\description{
Function to create a split train and validation samples
}
\examples{
library(tibble)
df <- tibble(i = 1:20)
split(df, ratio = .5)
}
