% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seer.R
\name{neophyte}
\alias{neophyte}
\title{Seer function to create forecasts}
\usage{
neophyte(df, index_var, y_var, x_vars = NULL, samples = list(strategy =
  "none"), models = list(list(algo = "auto.arima")), measure = "rmse",
  confidence_levels = c(0.8, 0.95), horizon = 1,
  forecast_xreg = NULL, backend = "sequential", user = NULL,
  uid = madutils::random_string("seer"), desc = "")
}
\arguments{
\item{df}{}

\item{index_var}{}

\item{y_var}{}

\item{x_vars}{}

\item{samples}{}

\item{models}{}

\item{measure}{}

\item{backend}{}
}
\description{
Seer function to create forecasts
}
